pyproject.toml
[project]
name = "flowapp"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Hodaka Mori", email = "hodaka.mori.office@gmail.com" }
]
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
    "barfi[streamlit]>=1.0.0",
    "rdkit>=2024.3.6",
    "scikit-learn>=1.6.0",
    "plotly>=5.24.1",
    "py3dmol>=2.4.2",
    "stmol>=0.0.9",
    "ipython-genutils>=0.2.0",
    "pip>=24.3.1",
    "grpcio>=1.68.1",
    "numpy==1.26.4",
    "altair>=5.5.0",
    "ydata-profiling>=4.12.1",
]


[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.7.1",
    "isort>=5.13.2",
    "black>=24.10.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.1",
    "ipykernel>=6.29.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["flowapp"]

[tool.ruff]
line-length = 88
exclude = [
    "*_pb2.py",
    "*_pb2_grpc.py",
]

[tool.ruff.lint]
extend-select = [
    "F",           # pyflakes
    # "W",           # pycodestyle-warnings
    "E",           # pycodestyle-errors
    # "C90",         # mccabe
    "I",           # isort
    # "N",           # pep8-naming
    "D",           # pydocstyle
    # "UP",          # pyupgrade
    # "YTT",         # flake8-2020
    # "ANN",         # flake8-annotations
    # "S",           # flake8-bandit
    # "BLE",         # flake8-blind-except
    # "FBT",         # flake8-boolean-trap
    # "B",           # flake8-bugbear
    # "A",           # flake8-builtins
    # "COM",         # flake8-comma
    # "C4",          # flake8-comprehensions
    # "T10",         # flake8-debugger
    # "EM",          # flake8-errmsg
    # "FA",          # flake8-future-annotations
    # "ISC",         # flake8-implicit-str-concat
    # "ICN",         # flake8-import-conventions
    # "G",           # flake8-logging-format
    # "PIE",         # flake8-pie
    # "T20",         # flake8-print
    # "PYI",         # flake8-pyi
    # "PT",          # flake8-pytest-style
    # "Q",           # flake8-quotes
    # "RSE",         # flake8-raise
    # "RET",         # flake8-return
    # "SLF",         # flake8-self
    # "SIM",         # flake8-simplify
    # "TID",         # flake8-tidy-imports
    # "TCH",         # flake8-type-checking
    # "ARG",         # flake8-unused-arguments
    # "PTH",         # flake8-use-pathlib
    # "TD",          # flake8-todos
    # "ERA",         # eradicate
    # "PD",          # pandas-vet
    # "PGH",         # pygrep-hooks
    # "PL",          # pylint
    # "TRY",         # tryceratops
    # "FLY",         # flynt
    # "NPY",         # NumPy
    # "PERF",        # perflint
    # "FURB",        # refurb
    # "RUF",         # Ruff-specific
]

extend-ignore = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D203",  # Conflict with D211
    "D205",
    "D213",  # Conflict with D212
]

[tool.mypy]
disable_error_code = ["import-untyped"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rdkit.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "psi4.*"
ignore_missing_imports = true
